[Refactor/#11] 검색화면 로직 리팩토링
- RxCocoa를 이용한 테이블뷰 데이터 바인딩
