+++ 34th-Sopt-Assignment/34th-Sopt-Assignment/Presentation/Find/View/BoxOfficeCell.swift
@@ -74,4 +74,3 @@ private extension BoxOfficeCell {
         }
     }
 }
-
