--- 34th-Sopt-Assignment/34th-Sopt-Assignment/Presentation/Find/ViewModel/Interface/FindViewModel.swift
@@ -1,18 +1,0 @@
-//
-//  FindViewModel.swift
-//  34th-Sopt-Assignment
-//
-//  Created by 김진웅 on 5/29/24.
-//
-
-import RxSwift
-
-protocol FindViewModelInput {
-    func viewDidLoad()
-}
-
-protocol FindViewModelOutput {
-    var isViewDidLoad: Observable<[DailyBoxOfficeList]> { get }
-}
-
-typealias FindViewModel = FindViewModelInput & FindViewModelOutput
